test(sdk): move generator tests under tests/
